requirements.txt
# Core
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings>=2.4.0

# Browser automation
playwright==1.58.0

# AI / LLM
google-genai==1.66.0
pillow==12.1.1

# Environment
python-dotenv==1.2.2
